pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cell2image"
version = "0.1.0"
description = "Toolbox for converting cell data to images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "vtk",
    "pillow",
    "matplotlib",
    "numba",
]
